# PERF: speed up MPLPlot._apply_axis_properties by skipping expensive tick enumeration where feasible